pyproject.toml
[tool.poetry]
name = "tietoa-slack-notifications"
version = "0.1.0"
description = ""
authors = ["Ville <ville.reima@tietoa.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
arrow = "^1.3.0"
slack-sdk = "^3.26.2"
loguru = "^0.7.2"
httpx = "^0.26.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
